update: 再利用性のないパーツをページファイルに移動
