+++ components/index.ts
@@ -1,6 +1,3 @@
 export { MenuItemIcon } from "./MenuItemIcon";
 export { MenuCard } from "./MenuCard";
 export { AddMenuCard } from "./AddMenuCard";
-export { HomeHead } from "./HomeHead";
-export { HomeEarnigs } from "./HomeEarnings";
-export { HomeGuide } from "./HomeGuide";
